Minor bug fix
If you were to attempt to send a transaction, with a total balance of 0 in the wallet, and then proceed to click either 25% or 50% then the sending balance would amount to `18,446,744,073,709.550616`. I've changed the percentage logic to just divide the balance - fee by 0.25 and 0.5 which when tested, will show a balance of `0.00` after pressing said buttons.